// interactive-map-back-office/src/models/MapMarkerModels/MapMarker.update.ts
export default interface MapMarkerUpdate {
    mapMarkerKey?: string;
    mapMarkerOrder?: number;
    mapMarkerName?: string;
    mapMarkerPositionX?: number;
    mapMarkerPositionY?: number;
    mapMarkerLocked?: boolean;
    mapMarkerTitle?: string;
    mapMarkerTitleDisplayType?: string;
    mapMarkerDescription?: string;
    mapMarkerPitchedUnitSelection?: string;
    mapMarkerButtonEnabled?: boolean;
    mapMarkerButtonLabel?: string;
    mapMarkerButtonURL?: string;
    mapMarkerImagesEnabled?: boolean;
    mapMarkerImages?: string[];
}